chore: changed services to interfaces
closes #13
