+++ src/main/java/msb/shortcut/master/service/AuditEventService.java
@@ -1,49 +1,16 @@
 package msb.shortcut.master.service;
 
-import msb.shortcut.master.config.audit.AuditEventConverter;
-import msb.shortcut.master.repository.PersistenceAuditEventRepository;
 import org.springframework.boot.actuate.audit.AuditEvent;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
 import java.util.Optional;
 
-/**
- * Service for managing audit events.
- * <p>
- * This is the default implementation to support SpringBoot Actuator AuditEventRepository
- */
-@Service
-@Transactional
-public class AuditEventService {
+public interface AuditEventService {
+    Page<AuditEvent> findAll(Pageable pageable);
 
-    private final PersistenceAuditEventRepository persistenceAuditEventRepository;
+    Page<AuditEvent> findByDates(Instant fromDate, Instant toDate, Pageable pageable);
 
-    private final AuditEventConverter auditEventConverter;
-
-    public AuditEventService(
-        PersistenceAuditEventRepository persistenceAuditEventRepository,
-        AuditEventConverter auditEventConverter) {
-
-        this.persistenceAuditEventRepository = persistenceAuditEventRepository;
-        this.auditEventConverter = auditEventConverter;
-    }
-
-    public Page<AuditEvent> findAll(Pageable pageable) {
-        return persistenceAuditEventRepository.findAll(pageable)
-            .map(auditEventConverter::convertToAuditEvent);
-    }
-
-    public Page<AuditEvent> findByDates(Instant fromDate, Instant toDate, Pageable pageable) {
-        return persistenceAuditEventRepository.findAllByAuditEventDateBetween(fromDate, toDate, pageable)
-            .map(auditEventConverter::convertToAuditEvent);
-    }
-
-    public Optional<AuditEvent> find(Long id) {
-        return Optional.ofNullable(persistenceAuditEventRepository.findOne(id)).map
-            (auditEventConverter::convertToAuditEvent);
-    }
+    Optional<AuditEvent> find(Long id);
 }
